chore: Move MessagePreview to common folder
